% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{checkCache}
\alias{checkCache}
\title{Check Cache vs Current}
\usage{
checkCache(overwrite = TRUE, ...)
}
\arguments{
\item{overwrite}{set to false to compare cache to file and return data}

\item{...}{Pass alternate \code{lineages_url} and \code{aliases_url} to \code{updateHierarchies}}
}
\value{
a list with three or more elements
\itemize{
\item lineages list with fixed aliases
\item hierarchy list
\itemize{
\item a Node class network object describing the hierarchy of lineages
\item a table of so-called orphan lineages that call on non-existent parents
}
\item file.info output showing most recent update time for cache
\item new lineages between cache and update
\item new aliases between cache and update
}
}
\description{
Checks for cache file, makes one if it doesn't exist, updates if it's out of
date, and loads it otherwise. Returns a list with the cached values. Option
to skip overwrite of cache and just return lineages. Custom target files for
\code{updateHierarchies()} can be supplied using \code{aliases_url} and \code{lineages_url}
}
