% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{constellationToTable}
\alias{constellationToTable}
\title{Parse Website Constellations from}
\usage{
constellationToTable(x)
}
\arguments{
\item{x}{location of cloned constellations root dir}
}
\value{
a table of constellation mutations with WHO label, variants, and mutations
}
\description{
https://github.com/cov-lineages/constellations/
}
