% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{expandChildren}
\alias{expandChildren}
\title{Get Descendants of a Table of Lineages}
\usage{
expandChildren(x, network, unique = TRUE)
}
\arguments{
\item{x}{Table of lineages (lineage column necessary)}

\item{network}{Network produced by \code{fromDFtoNetwork()}}

\item{unique}{Each child only has one parent}
}
\value{
Same table with all children added in child column
}
\description{
Using the covariants variants table or other and the network provided by
updateHierarchy or checkChache
}
