% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{fromDFtoNetwork}
\alias{fromDFtoNetwork}
\title{Build Lineage Node Network from Dataframe}
\usage{
fromDFtoNetwork(x)
}
\arguments{
\item{x}{Dataframe from \code{updateHierarchies()}}
}
\value{
a list with two elements
\itemize{
\item a Node class network object describing the hierarchy of lineages
\item a table of so-called orphan lineages that call on non-existent parents
}
}
\description{
Take a dataframe of parent child relationships and create a node network
that can be queried for relationships between them. Also returns a table of
orphans missing parents as these break the network
}
