% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{getVariants}
\alias{getVariants}
\title{Get CoVariants.org Tracked Variants}
\usage{
getVariants(
  variants_url =
    "https://raw.githubusercontent.com/hodcroftlab/covariants/master/web/data/nameTable.json"
)
}
\arguments{
\item{variants_url}{Point to a different version if necessary}
}
\value{
table of lineage, clade, who (designation)
}
\description{
Grab tracked variants table and a cleaned up tibble version of it.
}
