% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{makeWatchlist}
\alias{makeWatchlist}
\title{Make a Watchlist Appropriate to Augur}
\usage{
makeWatchlist(x)
}
\arguments{
\item{x}{Vector of lineage names}
}
\value{
A YAML ready array
}
\description{
Uses a character vector of lineage names to output a yaml array for Augur.
Vector can be manually produced or with \code{getDescendants()}
}
