% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{sortLineages}
\alias{sortLineages}
\title{Sort Lineages}
\usage{
sortLineages(x, column = "lineage")
}
\arguments{
\item{x}{Table with a lineage column to sort by}

\item{column}{Name of lineage column if not 'lineage'}
}
\value{
Table arranges by lineages in correct order
}
\description{
Sort lineages properly into the order defined by their sublineages
eg. A < A.1 < A.2 < A.2.1 < A.2.2 < A.2.3 < A.2.4 < A.2.5 < A.2.5.1
rather than: A.1 < A.10 < A.11 < A.12 < A.13 < A.14
}
