% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{uniqueListElements}
\alias{uniqueListElements}
\title{Find Unique List Elements}
\usage{
uniqueListElements(list)
}
\arguments{
\item{list}{sorted list of vectors}
}
\value{
same list where list elements have non-overlapping contents
}
\description{
Find Unique List Elements
}
