% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{updateHierarchies}
\alias{updateHierarchies}
\title{Update PANGO Hierarchies from Github Files}
\usage{
updateHierarchies(
  aliases_url =
    "https://raw.githubusercontent.com/cov-lineages/pango-designation/master/pango_designation/alias_key.json",
  lineages_url =
    "https://raw.githubusercontent.com/cov-lineages/pango-designation/master/lineage_notes.txt"
)
}
\arguments{
\item{aliases_url}{cov-lineages alias_key.json}

\item{lineages_url}{cov-lineages lineages_notes.txt}
}
\value{
dataframe of parent/lineage/children in long format
}
\description{
Cov-lineages curates a file that contains all lineages and one that contains
all aliases. Use modified code from Julie Chen to build hierarchy table.
}
