% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{checkCache}
\alias{checkCache}
\title{Check Cache vs Current}
\usage{
checkCache(overwrite = TRUE, network = FALSE, ...)
}
\arguments{
\item{overwrite}{set to false to compare cache to file and return data without writing}

\item{...}{Pass alternate \code{lineages_url} and \code{aliases_url} to \code{updateHierarchies}}

\item{output}{logical, TRUE for network (produced by fromDFtoNetwork) FALSE for table}
}
\value{
depending on the value of output
\itemize{
\item table of lineage hierarchies with fixed aliases
\item a Node class network object describing the hierarchy of lineages
}
}
\description{
Checks for cache file, makes one if it doesn't exist, updates if it's out of
date, and loads it otherwise. Returns a list with the cached values. Option
to skip overwrite of cache and just return lineages. Custom target files for
\code{updateHierarchies()} can be supplied using \code{aliases_url} and \code{lineages_url}
}
