% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{pullAncestors}
\alias{pullAncestors}
\title{Pull Ancestors from a Specific Lineage}
\usage{
pullAncestors(x, node_name, table = FALSE, vector = FALSE)
}
\arguments{
\item{x}{node network produce by \code{fromDFtoNetwork()}}

\item{node_name}{Lineage of interest ie. "BA.1"}

\item{table}{logical, return subtree or dataframe of level/lineage}

\item{vector}{logical, return vector  of lineage names}
}
\value{
one of two options depending on table param
\itemize{
\item a Node class network object of the nodes leading to node_name
\item a table from the node network with level and lineage names
\item a vector of lineage names
}
}
\description{
Get a lineages ancestors out of a \code{fromDFtoNetwork()}node network.
Vector being TRUE supersedes table being TRUE.
Default return node network.
}
